Fix keda-metrics-apiserver port
In our 2.11.2 release we retired the old 9022 metrics port, which was
superseded by port 8080. We updated the service with the port name
"metrics", but it looks like we missed updating the ServiceMonitor,
which still had it listed as "http" -- resulting in the prometheus
target not working.

This updates the ServiceMonitor so the port name matches the service,
and the tubes are properly connected and "metrics" will be a valid
prometheus target for keda-metrics-apiserver.

Signed-off-by: John Kyros <jkyros@redhat.com>