Make functionality tests wait for changes
Previously the functionality tests that tested the log level changes
were not waiting for rollouts to happen, so they were effectively not
testing that the logLevel and audit flags were properly set on the
deployment.

Recent releases have changed that timing so these tests have started to
fail.

This changes the test such that they now wait for controller to process
each of the changes (or not process each of the changes, depending on
the test case). It does this by adding an additional deployment
annotation for each test case, so the test case knows that the
controller has processed the test case and it can check the values.

This probably needs a more complete refactor at some point, but this at
least makes the test dependable.

The audit config tests don't work perfectly either, but that is harder
to solve because the functionality they are testing is currently kind of
broken, i.e. you can't currently deconfigure audit, so you can't ever
reset the test to a "clean" state to try the next test case. For now I
made them work by having them sequentially add the args without resetting.

Signed-off-by: John Kyros <jkyros@redhat.com>