Set root file system to read-only to match operand
The operand's root is set to read only, but the operator isn't, which is
causing things like the trivvy scanner to complain about it as "severity
high" misconfiguration.

This just changes the readOnlyRootFilesystem setting to true in the pod
templates (both the raw manifests and the CSV) to alleviate any possible
security issue the writeable root fs may cause. It should have no impact
whatsoever on the operation of the operator.

Signed-off-by: John Kyros <jkyros@redhat.com>