Fix volumeMount replacement, add tests
When we added volume mount overrides, it came with a bug such that it
would completely replace all the volume mounts, disregarding the
merge/add logic. This works fine for the most part, except in cases
where someone has specified caConfigMaps, in which case they just get
completely blown out of the list.

This removes the erroneous line and adds tests to enforce the behavior.
The removal of the line does not break existing tests or behavior.

Signed-off-by: John Kyros <jkyros@redhat.com>