Update golangci-lint to support v2 export data format
Our golangci-lint was too old to process newer go code, it couldn't
properly read the imports. This moves us to the most recent v1 version.

There is a v2 version now that we need to move to, but we need to do the
migration, and the linters are all a little different, so it's
nitpicking a bunch of issues that have been in there forever. We'll come
back and do that migration as a separate PR.

Signed-off-by: John Kyros <jkyros@redhat.com>